+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add filter for healthcheck endpoints. ([#2128](https://github.com/getsentry/relay/pull/2762))
+
 **Internal**:
 
 - Add `txNameReady` flag to project config. ([#2128](https://github.com/getsentry/relay/pull/2128))