+++ py/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add filter for healthcheck endpoints. ([#2128](https://github.com/getsentry/relay/pull/2762))
+
 ## 0.8.23
 
 - Add `txNameReady` flag to project config. ([#2128](https://github.com/getsentry/relay/pull/2128))